Updating Qt support code to handle the various versions of Qt.
We do a more exacting search for versions in the 5.6.x, 5.7.x, 5.8.x and
5.9.x range by using some assumptions about the maximum version number
that has ever been released for that minor version of Qt.

Also place a FATAL_ERROR in the Qt 5.8.x code that says if a developer
is on Windows and tries Qt 5.8 we FATAL_ERROR on that due to a bug in Qt
that stops Q_INTERFACES being parsed correctly when using CMake.